+++ AVR/accel.h
@@ -163,7 +163,7 @@ typedef enum
 void initI2C(void);
 void bno055SetReg(uint8_t reg, uint8_t data);
 uint8_t bno055ReadReg(uint8_t reg);
-void reportData(uint8_t x, uint8_t y, uint8_t z);
+void reportData(int16_t x, int16_t y, int16_t z);
 float tilt_angle(int8_t x, int8_t y);
 
 #endif
